Public Beta Version 4.46
Fixed an issue where the resource was requesting the objects from the server every single time a resource was started, instead of just one time, when the Fire Alarm resource was started. This issued caused intense server strain, resulting in `Rage Crashes`, and texture issues on large servers (i.e. servers with hundreds of resources).